Rename module `base` to `fields`
